react/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
snowflake-connector-python[secure-local-storage]>=3.12.0
sse-starlette>=2.0.0
pydantic>=2.0.0
cryptography>=42.0.0
